4.deployment_dev/streamlit_KG_textimg/streamlitKG-image-deployment/requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
matplotlib
ipython
Pillow
transformers==4.37.2
bitsandbytes==0.41.3
accelerate==0.25.0
neo4j
networkx
graphviz
plotly
torch
requests